Clean up files
- Fix Gradle
- Remove clutter
- Update MCP mappings
- Refactor to use io.github.jas777.ontherails instead of com.jas777.ontherails
- Change ModID to ontherails
