--- src/main/java/com/jas777/ontherails/Reference.java
@@ -1,14 +1,0 @@
-package com.jas777.ontherails;
-
-public class Reference {
-    
-    //Main mod info
-    public static final String MOD_ID = "otr";
-    public static final String NAME = "On The Rails";
-    public static final String VERSION = "0.1-earlyacces";
-    public static final String ACCEPTED_VERSIONS = "[1.11.2]";
-    
-    //Proxies
-    public static final String CLIENT_PROXY_CLASS = "com.jas777.ontherails.proxy.ClientProxy";
-    public static final String SERVER_PROXY_CLASS ="com.jas777.ontherails.proxy.ServerProxy";
-}
