--- src/main/java/com/jas777/ontherails/proxy/ClientProxy.java
@@ -1,5 +1,0 @@
-package com.jas777.ontherails.proxy;
-
-public class ClientProxy implements com.jas777.ontherails.proxy.CommonProxy {
-
-}
