--- src/main/java/com/jas777/ontherails/proxy/CommonProxy.java
@@ -1,5 +1,0 @@
-package com.jas777.ontherails.proxy;
-
-public interface CommonProxy {
-
-}
