--- src/main/java/com/jas777/ontherails/proxy/ServerProxy.java
@@ -1,5 +1,0 @@
-package com.jas777.ontherails.proxy;
-
-public class ServerProxy implements com.jas777.ontherails.proxy.CommonProxy{
-
-}
