+++ src/main/java/io/github/jas777/ontherails/OnTheRails.java
@@ -1,0 +1,45 @@
+package com.jas777.ontherails;
+
+import com.jas777.ontherails.proxy.CommonProxy;
+
+import net.minecraftforge.fml.common.Mod;
+import net.minecraftforge.fml.common.Mod.EventHandler;
+import net.minecraftforge.fml.common.Mod.Instance;
+import net.minecraftforge.fml.common.SidedProxy;
+import net.minecraftforge.fml.common.event.FMLInitializationEvent;
+import net.minecraftforge.fml.common.event.FMLPostInitializationEvent;
+import net.minecraftforge.fml.common.event.FMLPreInitializationEvent;
+
+//References
+
+@Mod(modid = Reference.MOD_ID, name = Reference.NAME, version = Reference.VERSION, acceptedMinecraftVersions = Reference.ACCEPTED_VERSIONS)
+public class OnTheRails {
+ // Sascha Test Note
+    @Instance
+    public static OnTheRails instance;
+    
+    // Proxy
+    
+    @SidedProxy(clientSide = Reference.CLIENT_PROXY_CLASS, serverSide = Reference.SERVER_PROXY_CLASS)
+    public static CommonProxy proxy;
+    
+    // Initialization(s)
+    
+    @EventHandler
+    public void preInit(FMLPreInitializationEvent event)
+    {
+        
+    }
+    
+    @EventHandler
+    public void init(FMLInitializationEvent event)
+    {
+        
+    }
+    
+    @EventHandler
+    public void postInit(FMLPostInitializationEvent event)
+    {
+        
+    }
+}
