+++ src/main/java/io/github/jas777/ontherails/proxy/CommonProxy.java
@@ -1,0 +1,5 @@
+package com.jas777.ontherails.proxy;
+
+public interface CommonProxy {
+
+}
